Act086ProductItemAdapter  - Renomeado paa Act086MaterialItemAdapter  - Refactor nas antigas rferencias product item  - Add properties inReadonly e repassado para o vh @#$ Act086ProductItemBinding e Act086ProductItemVH  - Refector para Act086MaterialItemBinding e Act086MaterialItemVH  - Act086ProductItemVH add properties inReadonly e aplicado nas views. Caso seja readonly, remove os clique e o icone de lixeira @$ Act086MaterialItem  - Modificado data type do campo productQty para float @#$ GeOsDeviceItem  - Modificado classe para ser usada no envio do form, add notation @Expose nos campos que precisam ser visiveis e add SerializedName na lista de material. @#$ GeOsDeviceMaterial  - Modificado classe para ser usada no envio do form, add notation @Expose nos campos que precisam ser visiveis @#$ TSave_Env  - Add lista de device items ao envio @#$ WS_Save  - Add lista de GeOsDeviceItem sob o nome de form_items  - Add form_items no obj de envio  - Modificado metodos processPendingToken e processNewToken para resgatarem tb o itens do device da o.s @#$ GeOsDeviceItem_Sql_003  - Query que lista item com fotos para serem inseidas na listagem de geFiles @#$ Sql_WS_Save_Device_Item_001  - Query que seleciona os device item da os relacionada ao form @#$ Act011_Main  - Criado metodo addGeOsDeviceItemPhotosIntoFiles que caso o form seja do tipo o.s, insere as imagens dos device_items na tabela ge_files  - Modificado metodo onInspectionSelected add param com status do form no bundle de device. Info sera usada para deixar o item como readonly @#$ Act086VerificationFrg  - Criado fun applyReadonlyUi que aplica o readonly nas views e adapters  - Modificado fun getnewIntance para receber e armazenar no args a proprieade inReadOnly  - Add passagem desse novo param no construtor do Act086MaterialItemAdapter  - Modificado fun loadMaterialListFromDb, add chamada fun updateMaterialLabelCount para atualizar contador no lbl  - Corrigido logica do setOnCheckedChangeListener, pois estava analisando se lista de photo não era vazia. No casso, deve ser analisado somnete a lista de material  - Modificado fun callCameraAct , modificado param PENABLED, para deixar desabilitado caso readonly  - Modificado onPause para pular o save em caso de readonly  - Removido toInto na geração do obj Act086MaterialItem na fun buildAdapterMaterialFragList @#$ Act086Main.kt  - Add resgate do status do form no bundle para alimentar a var readOnly via fun defineReadOnlyByStatus  - Atualizado label do campo de data para ficarem igual ao da listagem na fun getAlertDateLbl. Tb corrigido metodo de conversão da diferença das datas para usar o toDay e não to hour @#$ Act086ProductEditDialog.kt  - Modificado o get do valor do mket para toFloat e não to int

+++ app/src/main/java/com/namoadigital/prj001/model/Act086MaterialItem.kt
@@ -7,7 +7,7 @@ class Act086MaterialItem(
     productId: String,
     productDesc: String,
     productUnit: String,
-    productQty: Int = 0,
+    productQty: Float = 0f,
     creationMs: Long = 0
 ):Serializable{
     var productCode: Int = -1
@@ -18,7 +18,7 @@ class Act086MaterialItem(
         private set
     var productUnit: String = ""
         private set
-    var productQty: Int = 0
+    var productQty: Float = 0f
 
     var creationMs: Long = 0
         private set
@@ -32,5 +32,5 @@ class Act086MaterialItem(
         this.creationMs = creationMs
     }
 
-    fun getFormttedQty() = if(productQty > 0) "$productQty $productUnit" else ""
+    fun getFormttedQty() = if(productQty > 0f) "$productQty $productUnit" else ""
 }
